test: UI Test 패키지 분리
